chore: split rules and endpoints on diff layers
